+++ changelog.md
@@ -4,7 +4,7 @@
 
 ### Version 4.0.11
 
-* `cordapp`: Ensure that jitpack versions of dependencies are also excluded from the "fat jar" [CORDA-1301].
+* `quasar-utils`: Add quasar-core to cordaRuntime. This ensures that it is excluded from "fat jar" CorDapps [CORDA-1301].
 
 ### Version 4.0.10